Initialize dialer late, use tcpMonitor for dialer construction
